requirements.txt
asgiref==3.9.1
Django==5.2.4
django-ckeditor==6.7.3
django-debug-toolbar==6.0.0
django-filter==25.1
django-js-asset==3.1.2
django-querycount==0.8.3
django-resized==1.0.3
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
Markdown==3.8.2
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.14.1
Unidecode==1.4.0
uritemplate==4.2.0
